@@ -1,18 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
 using NLog.Extensions.Logging;
-using NLog.Web;
 
 namespace ISTC.FiveStage.Technology.Logging.Log_01
 {
@@ -28,13 +21,13 @@ namespace ISTC.FiveStage.Technology.Logging.Log_01
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
             services.AddLogging();
 
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, ILoggerFactory loggerFactory)
         {
             if (env.IsDevelopment())
             {
@@ -46,11 +39,17 @@ namespace ISTC.FiveStage.Technology.Logging.Log_01
                 app.UseHsts();
             }
            
+#pragma warning disable 618
             loggerFactory.AddNLog();
-            env.ConfigureNLog("nlog.config");
+#pragma warning restore 618
+            //env.ConfigureNLog("nlog.config");
 
+
+            app.UseRouting();
             app.UseHttpsRedirection();
-            app.UseMvc();
+
+            app.UseAuthorization();
+            app.UseEndpoints((x) => x.MapControllers());
         }
     }
 }
